fixed a wcs rotation issue
Former-commit-id: 612f5fc3cebe1672191cef91cf32840128cc7345
Former-commit-id: 771292be946d266e397c64a0c611040a4ab1f873

+++ java/src/edu/caltech/ipac/firefly/server/visualize/VisContext.java
@@ -94,13 +94,14 @@ public class VisContext {
 
     static {
         boolean speed=AppProperties.getBooleanProperty("visualize.fits.OptimizeForSpeed",true);
-        if (speed) {
-            purger= new OptimizeForSpeedPurger();
-        }
-        else {
-            purger= new OptimizeForMemoryPurger();
-
-        }
+//        if (speed) {
+//            purger= new OptimizeForSpeedPurger();
+//        }
+//        else {
+//            purger= new OptimizeForMemoryPurger();
+//
+//        }
+        purger=  null;
 
         init();
     }
@@ -420,7 +421,7 @@ public class VisContext {
     }
 
     public static void purgeOtherPlots(PlotState state) {
-        purger.purgeOtherPlots(state);
+        if (purger!=null) purger.purgeOtherPlots(state);
     }
